last framework compile

+++ app/application/example/src/main/java/com/igumnov/common/entities/Cl.java
@@ -15,5 +15,5 @@ public class Cl {
     @Field(fieldName = "gender" , autoIncremental = false)
     private String sex;
     @Id(autoIncremental = true, fieldName = "id")
-    private BigDecimal id;
+    private Integer id;
 }
